% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheese.R
\docType{data}
\name{cheese}
\alias{cheese}
\title{loads PNNL dataframes}
\format{
dataframe
}
\source{
cheese data
}
\description{
This function loads three PNNL dataframes: dfgroup_long, ptw and res.
}
\examples{
data(cheese)
head(c12)
head(f12)
head(t12)
}
