% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes__.R
\docType{package}
\name{color_palettes__}
\alias{-package}
\alias{color_palettes__}
\title{Color palettes}
\usage{
color_palettes__(name)
}
\arguments{
\item{name}{A string with two options: \code{"graytone"} or \code{"darkwarm"}.}
}
\description{
An internal function for built-in color palettes.
}
\examples{
color_palettes__("graytone")
color_palettes__("darkwarm")

}
