% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_range.R
\docType{package}
\name{convert_range}
\alias{convert_range}
\title{Convert range}
\usage{
convert_range(x)
}
\arguments{
\item{x}{A vector of numbers.}
}
\value{
A vector calculated by (x - min(x)) / (max(x) - min(x))
}
\description{
An internal function for range conversion. For each element in x, it returns the distance to the minimal values and divided by the range.
}
\examples{
convert_range(5:10)

}
