% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.color__.R
\docType{package}
\name{extend.color__}
\alias{extend.color__}
\title{Extend color palettes}
\usage{
extend.color__(n, colors, extendby = 1, alpha = 1, ...)
}
\arguments{
\item{n}{A number. The length of feature vector that you want to match with the colors.}

\item{colors}{A vector of colors (finite).}

\item{extendby}{A number to select from 1, 2, 3, or 4, representing four distinct auto-fill schemes.}

\item{alpha}{A number to select from range \link{0,1}.}
}
\description{
An internal function designed to pair vectors with color palettes and automatically generate colors for longer vectors that cannot find a match.
}
