% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gAnnotation.R
\docType{package}
\name{gAnnotation}
\alias{gAnnotation}
\title{gAnnotation}
\usage{
gAnnotation(data, first_order, second_order, identifier)
}
\arguments{
\item{data}{A data frame generated by gData().}

\item{ncluster}{A number of clusters.}

\item{iter.max}{A number of the maximum iterations allowed in k-means.}

\item{nstart}{A number of random attempts of generating initial
configurations. The k-means algorithm will choose the best one among these
attempts. For larger data, 'nstart' can be set lower or just set to 1.}

\item{elbow.max}{A number of the maximum value of x-axis. It should be larger
than the expected \code{ncluster} and smaller than the sample size.}
}
\value{
A list of 2:
}
\description{
\code{gClusters} returns the
}
\details{
This function could
}
\examples{
data(FuncExample)
}
