% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gClusters.R
\docType{package}
\name{gClusters}
\alias{gClusters}
\title{gClusters}
\usage{
gClusters(data, ncluster = 20, elbow.max = 50, ...)
}
\arguments{
\item{data}{A data frame generated by gData().}

\item{ncluster}{A number of clusters.}

\item{elbow.max}{A number of the maximum value of x-axis. It should be larger
than the expected \code{ncluster} and smaller than the sample size.}

\item{iter.max}{A number of the maximum iterations allowed in k-means.}

\item{nstart}{A number of random attempts of generating initial
configurations. The k-means algorithm will choose the best one among these
attempts. For larger data, 'nstart' can be set lower or just set to 1.}
}
\value{
A list of 2:
}
\description{
\code{gClusters} returns the clusters generated by k-means and yield
an elbow plot as a way of finding the optimal parameter.
}
\details{
This function could be ran without any parameters except \code{data} at
the very beginning. After that, users are expected to tune the values
according to the clusters and elbow plot. To determine the optimal number
of \code{ncluster}, select the value at where the graph has a rapid change. You
may also increase the elbow.max if the data is very large or complicated.
}
\examples{
data(FuncExample)
head(a)
reslist <- gClusters(a)
# k-means result
reslist[[1]]
# elbow plot
reslist[[2]]

}
