% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gClusters.R
\docType{package}
\name{gClusters}
\alias{gClusters}
\title{Generate clusters}
\usage{
gClusters(data, ncluster = 20, elbow.max = 50, ...)
}
\arguments{
\item{data}{A scaled data.frame that contain variables \code{ID}, \code{value on time_1}, ..., \code{value on time_k}, and \code{type} for extracting patterns across the time. See also \code{pre_process()}.}

\item{ncluster}{A number of clusters. It is related to the complexity of information in the network: When choosing the \code{ncluster}, we suggest thinking about how many nodes you are about to show on the visualization and how representative you want for each clustered pattern.}

\item{elbow.max}{A number of the maximum value of x-axis for the elbow method plot. It should be larger than the expected \code{ncluster} and smaller than the sample size.}

\item{iter.max}{A number of the maximum iterations allowed in k-means. See also \code{stats::kmeans}.}

\item{nstart}{A number of random attempts of generating initial configurations. The k-means algorithm will choose the best one among these
attempts. For larger data, 'nstart' can be set lower or just set to 1. See also \code{stats::kmeans}.}
}
\value{
This function return a list of 2 elements: a k-means cluster result and an elbow method plot.
}
\description{
\code{gClusters()} returns the clusters generated by k-means and yield
an elbow plot as a way of finding the optimal parameter.
}
\details{
To determine the optimal number of clusters (\code{ncluster}), it is advised to closely examine the elbow plot and identify the point on the graph where a substantial change or 'elbow' occurs. This is often indicative of the most suitable cluster count.
In cases where your dataset is extensive or intricate, you might consider increasing the value of \code{elbow.max} to ensure a more comprehensive exploration of potential cluster counts. This can help in achieving more accurate and meaningful results, especially when working with larger or more complex datasets. This function can be executed with only the \code{data} parameter at the outset. However, to achieve the best clustering results, further adjustments are recommended. After the initial run, users are expected to adjust the function's parameters based on the clustering outcomes and the elbow plot analysis.
}
\examples{
data(test_data)
reslist <- gClusters(test_data_processed)
# k-means result
reslist[[1]]
# elbow plot
reslist[[2]]

}
