% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDashboard.R
\docType{package}
\name{gDashboard}
\alias{gDashboard}
\title{gDashboard}
\source{
scale()
}
\usage{
gDashboard(
  data,
  clusters,
  pathway,
  networkres,
  dashboardtitle = "MolPad Dashboard",
  id_type = "KEGG"
)
}
\arguments{
\item{data}{The output of \code{pre_process()}}

\item{clusters}{The output of \code{gClusters()}}

\item{pathway}{A dataframe. The columns must include "ID", "taxonomic.scope",
and "Pathway".}

\item{networkres}{The output of \code{gNetwork}}

\item{dashboardtitle}{A string.}

\item{annotation}{A string. Choose a column in your dataset to add links for.}
}
\description{
Run shiny dashboard.
}
\examples{
data(FuncExample)
gDashboard(a, b, pathway, networkres, dashboardtitle = "My Title")
}
