% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMainData.R
\docType{package}
\name{gMainData}
\alias{gMainData}
\title{gMainData}
\source{
scale()
}
\usage{
gMainData(data, clusters, pathway)
}
\arguments{
\item{data}{The output of \code{pre_process()}}

\item{clusters}{The output of \code{gClusters()}}

\item{pathway}{he output of \code{gPathway()}}
}
\description{
This function generate a main data-input for the dashboard.
}
\examples{
data(FuncExample)
# inputs
head(a)
head(b)
head(ptw)
# generate dataset
dfgroup_long <- gMainData(a, b, ptw)
head(dfgroup_long)
}
