% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gNetwork.R
\docType{package}
\name{gNetwork}
\alias{gNetwork}
\title{Generate prediction network}
\usage{
gNetwork(clusters, ntop = 10, method = "randomforest")
}
\arguments{
\item{clusters}{A list of two outputs of \code{gClusters()}. The first element is the k-means result (See also \code{stats::kmeans}), while the other element is a plot and will be automatically omitted for ease of directly passing results to the function.}

\item{ntop}{A number you pick to set the top n predictors(default: top 10) for each feature as the important ones. This will be used to construct the network among the clustered patterns.}

\item{seed}{A number of random seed.}
}
\description{
\code{gNetwork()} generates a prediction network for each functional annotation. For every feature, all other features are considered as independent variables, and the top predictors are selected based on \%IncMSE.
}
\details{
We assess relationships between clusters using Mean Squared Error (MSE) changes resulting from random shuffling. \code{gNetwork()}'s output includes edge weights and node pairs, which are essential inputs for the dashboard.
}
\examples{
data(test_data)
networkres <- gNetwork(test_cluster, ntop = 3)
head(networkres)
gNetwork_view(networkres)

}
