% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gNetwork_view.R
\docType{package}
\name{gNetwork_view}
\alias{gNetwork_view}
\title{gNetwork_view}
\source{
scale()
}
\usage{
gNetwork_view(data)
}
\arguments{
\item{data}{The output of \code{gNetwork()}}
}
\description{
Generate a feature importance plot for the random forest result.
}
\examples{
data(FuncExample)
gNetwork_view(networkres)
}
