% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stackbar_plot.R
\docType{data}
\name{make_stackbar_plot}
\alias{make_stackbar_plot}
\title{Make stackbar plot}
\source{
ggplot2::geom_bar
}
\usage{
make_stackbar_plot(dfgroup_long, selected_groups, selected_taxa)
}
\description{
Generate the stackbar plot for the dashboard.
}
\examples{
data(test_data)
make_stackbar_plot(test_maindata, "Group_5", c("hormonal proteins","structural proteins","enzymes","storage proteins","antibodies","transport proteins"))

}
