% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_the _table.R
\docType{package}
\name{make_the_table}
\alias{make_the_table}
\title{Make table from brushed region}
\usage{
make_the_table(p)
}
\arguments{
\item{p}{ggplot output}
}
\description{
Generate the table of selected groups for the dashboard.
}
