% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_produce_lm__.R
\docType{package}
\name{mass_produce_lm__}
\alias{mass_produce_lm__}
\title{Mass produce linear model}
\usage{
mass_produce_lm__(data)
}
\arguments{
\item{data}{A dataframe.}
}
\description{
Take each column as dependent vairable at a time, and produce n linear models for n columns in a dataset.
}
\details{
This is an internal function designed to automatically generate a list of functions for regression. Each column is considered a response variable against all other columns. The resulting functions can be utilized as inputs for random forest regression in \code{gNetwork()}.
}
