% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.color__.R
\docType{package}
\name{match.color__}
\alias{match.color__}
\title{Match color}
\usage{
match.color__(valist, mycolors, extendby = 5)
}
\arguments{
\item{valist}{A vector that you want to specify the colors for each element.}

\item{mycolors}{A vector of colors.}

\item{extendby}{A number to select from 1, 2, 3, 4, or 5, representing distinct auto-fill schemes.}
}
\description{
Match any vector with a finite list of colors.
}
\examples{
my_vector <- paste0("N_",1:10)
match.color__(my_vector,c("red","yellow","blue"))

}
