% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_database.R
\docType{package}
\name{match_database}
\alias{match_database}
\title{match database}
\format{
A dataframe and a weblink column name
}
\usage{
match_database(data, id_colname, id_type)
}
\description{
Generate the cleaned version of the pathway for the dashboard.
}
\examples{
data(stuff)
nao_ptw <- ptw_process(ptw, ko_term)
head(nao_ptw)
}
