% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_URL.R
\docType{package}
\name{paste_URL}
\alias{paste_URL}
\title{Paste URL}
\usage{
paste_URL(x, id_type)
}
\description{
Retrieve database IDs and associate them with their respective URLs.
}
\examples{
data(test_data)
paste_URL(test_annotations$GO_ID[1:4], id_type = "GO")

}
