% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\docType{package}
\name{pre_process}
\alias{pre_process}
\title{pre_process}
\usage{
pre_process(
  dataList,
  typenameList = NULL,
  replaceNA = TRUE,
  scale = TRUE,
  autoColName = "Day_"
)
}
\arguments{
\item{dataList}{A list of datasets. Note that for each dataset, the first columns should be ID and the rest columns are Time1, Time2, ...}

\item{typenameList}{A vector of dataset names.}

\item{scale}{Logical; if TRUE (default), scale the datasets by row. See also \link[pkg]{scale_by_row__}}

\item{autoColName}{A string; if it's not NULL (default), automatically set uniform column names for all the datasets.}

\item{replaceNa}{Logical; if TRUE (default), replace NA with 0.}
}
\description{
This function preprocess the data-inputs and automatically sets uniform format for future use.
}
\examples{
data(samples)
a <- pre_process(samples)
head(a, 10)
}
