% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptw_process.R
\docType{data}
\name{ptw_process}
\alias{ptw_process}
\title{process pathway}
\format{
A dataframe and a weblink column name
}
\usage{
ptw_process(ptw, ko_term)
}
\description{
Generate the cleaned version of the pathway for the dashboard.
}
\examples{
data(stuff)
nao_ptw <- ptw_process(ptw, ko_term)
head(nao_ptw)
}
