% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_for_make_functions.R
\docType{package}
\name{reshape_for_make_functions}
\alias{reshape_for_make_functions}
\title{reshape for make functions}
\source{
scale()
}
\usage{
reshape_for_make_functions(data, cluster, annotation, id_type)
}
\arguments{
\item{data}{The output of \code{pre_process()}}

\item{cluster}{The output of \code{gClusters()}}

\item{annotation}{he output of \code{gPathway()}}
}
\description{
This function
}
