% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_by_row__.R
\docType{package}
\name{scale_by_row__}
\alias{scale_by_row__}
\title{scale_by_row}
\format{
A data frame in which the first column is ID and the rest are
observations for different time points. Note that the first column should be ID
and thus will not be scaled, and the rest of the columns require dbl format.
}
\source{
scale()
}
\usage{
scale_by_row__(data)
}
\description{
This function scales the values in the rows for all non-ID columns.
}
\examples{
data(fungal)
scale_by_row__(fungal[[3]])
}
