% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stuff}
\alias{stuff}
\title{loads PNNL dataframes}
\format{
dataframe
}
\source{
Generated from PNNL
}
\description{
This function loads three PNNL dataframes: dfgroup_long, ptw and res.
}
\examples{
data(stuff)
head(dfgroup_long)
head(ptw)
head(res)
}
