% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{loads PNNL dataframes}
\format{
dataframe
}
\source{
cheese data
}
\description{
This function loads three PNNL dataframes: dfgroup_long, ptw and res.
}
\examples{
data(test_data)
head(c12)
head(f12)
head(t12)
}
