% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_dataframe__.R
\docType{package}
\name{transpose_dataframe__}
\alias{transpose_dataframe__}
\title{transpose dataframe}
\format{
A data frame in which the first column is ID and the rest are observations for different time points.
}
\usage{
transpose_dataframe__(data)
}
\description{
This function transpose a given data frame, converting the first column to new column names
}
\examples{
data(fungal)
transpose_dataframe__(fungal[[3]])
}
