% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheese.R
\docType{data}
\name{cheese}
\alias{cheese}
\title{loads PNNL dataframes}
\format{
A data frame with 234 rows and 11 variables:
\describe{
\item{manufacturer}{manufacturer name}
\item{model}{model name}
\item{displ}{engine displacement, in litres}
\item{year}{year of manufacture}
\item{cyl}{number of cylinders}
\item{trans}{type of transmission}
\item{drv}{the type of drive train, where f = front-wheel drive, r = rear wheel drive, 4 = 4wd}
\item{cty}{city miles per gallon}
\item{hwy}{highway miles per gallon}
\item{fl}{fuel type}
\item{class}{"type" of car}
}

This function loads three PNNL dataframes: dfgroup_long, ptw and res.

dataframe
}
\source{
cheese data
}
\description{
loads PNNL dataframes
}
\examples{
data(cheese)
head(c12)
head(f12)
head(t12)
}
