% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungal_garden.R
\docType{data}
\name{fungal_garden}
\alias{fungal_garden}
\title{loads PNNL dataframes}
\format{
A data frame with 234 rows and 11 variables:
\describe{
\item{manufacturer}{manufacturer name}
\item{model}{model name}
\item{displ}{engine displacement, in litres}
\item{year}{year of manufacture}
\item{cyl}{number of cylinders}
\item{trans}{type of transmission}
\item{drv}{the type of drive train, where f = front-wheel drive, r = rear wheel drive, 4 = 4wd}
\item{cty}{city miles per gallon}
\item{hwy}{highway miles per gallon}
\item{fl}{fuel type}
\item{class}{"type" of car}
}
This function loads three PNNL dataframes: dfgroup_long, ptw and res.

dataframe
}
\source{
Generated from PNNL
}
\description{
loads PNNL dataframes
}
\examples{
data(fungal_garden)
head(dfgroup_long)
head(ptw)
head(res)
}
