% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gAnnotation.R
\docType{package}
\name{gAnnotation}
\alias{gAnnotation}
\title{Generate processed annotation}
\usage{
gAnnotation(data, first_order, second_order)
}
\arguments{
\item{data}{A data.frame containing annotations that should be used for describing the features we measured at each time point. It is also required to include \code{ID} and at least two categorical variables.}

\item{first_order}{A string. The name of one column of categorical variable in \code{data}.}

\item{second_order}{A string. The name of another column of categorical variable in \code{data}.}
}
\description{
\code{gAnnotation()} provides the standard input format for the dashboard, allowing users to select two columns as the primary factors they wish to visualize and describe in the data.
}
\details{
Guidelines for Selecting Annotations:
\itemize{
\item The first-order annotation is recommended to be functional, such as a pathway or functional system. This parameter will primarily serve the purpose of filtering out one network at a time when displaying the dashboard.
\item The second-order annotation is utilized to illustrate the composition of a first-order annotation using a bar plot. Therefore, it is better to set it as taxon, class label, etc.
}
}
\examples{
data(test_data)
test_annotations_processed <- gAnnotation(test_annotations,system,class)
head(test_annotations_processed)

}
