% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gAnnotation.R
\docType{package}
\name{gAnnotation}
\alias{gAnnotation}
\title{generate processed annotation}
\usage{
gAnnotation(data, first_order, second_order)
}
\arguments{
\item{data}{A dataframe.}

\item{first_order}{The name of one column of categorical variable in \code{data}.}

\item{second_order}{The name of another column of categorical variable in \code{data}.}
}
\description{
\code{gAnnotation()} returns the
}
\details{
This function could

Guidelines for Selecting Annotations:
*The first-order annotation is recommended to be functional, such as a pathway or functional system. This parameter will primarily serve the purpose of filtering out one network at a time when displaying the dashboard.
*The second-order annotation is utilized to illustrate the composition of a first-order annotation using a bar plot. Therefore, it is better to set it as taxon, class label, etc.
}
\examples{
data(test_data)
test_annotations_processed <- gAnnotation(test_annotations,system,class)
head(test_annotations_processed)

}
