% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDashboard.R
\docType{package}
\name{gDashboard}
\alias{gDashboard}
\title{gDashboard}
\source{
scale()
}
\usage{
gDashboard(
  data,
  cluster,
  annotation,
  networkres,
  dashboardtitle = "MolPad Dashboard",
  id_type = "KEGG"
)
}
\arguments{
\item{data}{The output of \code{pre_process()}}

\item{cluster}{The output of \code{gClusters()}}

\item{annotation}{A string. Choose a column in your dataset to add links for.}

\item{networkres}{The output of \code{gNetwork}}

\item{dashboardtitle}{A string.}
}
\description{
Run shiny dashboard.
}
\examples{
data(FuncExample)
gDashboard(a, b, annotation, networkres, dashboardtitle = "My Title")
}
