% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDashboard.R
\docType{package}
\name{gDashboard}
\alias{gDashboard}
\title{Generate shiny dashboard}
\usage{
gDashboard(
  data,
  cluster,
  annotation,
  networkres,
  dashboardtitle = "MolPad Dashboard",
  id_colname = NULL,
  id_type = NULL
)
}
\arguments{
\item{data}{The output of \code{pre_process()}:a scaled data.frame contain variables \code{ID}, \code{value on time_1}, ..., \code{value on time_k}, and \code{type}.}

\item{cluster}{The output of \code{gClusters()}: a list contains the result of k-means Cluster means, vectors and sum of squares.}

\item{annotation}{The output of \code{gAnnotation()}: a data.frame containing annotations for describing the features. Variables must include \code{ID}, \code{Pathway}, and \code{taxonomic.scope}. Note that NA is not permitted for these 3 variables.}

\item{networkres}{The output of \code{gNetwork()}: a data.frame with 4 variables \code{weight},\code{IncNodePurity},\code{var_names},\code{from}.}

\item{dashboardtitle}{A string of customized dashboard name.}

\item{id_colname}{A single string or a string vector. The column names in your annotation dataset where the vector contains external database IDs.}

\item{id_type}{A single string or a string vector. The corresponding database names for the above columns, must be chosen from "KEGG" and "GO".}
}
\description{
Once you have the outputs from the g-functions, you're ready to create your custom Molpad dashboard. Be sure to specify the web ID columns and their corresponding column names.
}
\details{
Please ensure that columns containing external database IDs adhere to the following standards:
\itemize{
\item KEGG ID: Begin with 'K' followed by 5 digits, for example, K05685 or K06671.
\item GO ID: Begin with 'GO:' followed by 7 digits, such as GO:0003674."
}
}
\examples{
\dontrun{
data(test_data)
gDashboard(test_data_processed,test_cluster,test_annotations_processed,test_network,id_colname = c("GO_ID","KEGG_ID"),id_type = c("GO","KEGG"))
}
}
