% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gNetwork.R
\docType{package}
\name{gNetwork}
\alias{gNetwork}
\title{gNetwork}
\source{
scale()
}
\usage{
gNetwork(clusters, ntop = 10, method = "randomforest")
}
\arguments{
\item{clusters}{The output of \code{gClusters()}}

\item{ntop}{A number.Pick the top n predictors for each feature.}

\item{seed}{A number of random seed.}
}
\description{
Generate random forest outputs for the Network.
}
\examples{
data(FuncExample)
networkres <- gNetwork(b, ntop = 3)
head(networkres)
gNetwork_view(networkres)

}
