% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_line_plot.R
\docType{data}
\name{make_line_plot}
\alias{make_line_plot}
\title{Make line plot}
\source{
ggplot2::geom_ribbon
}
\usage{
make_line_plot(dfgroup_long, selected_groups, selected_taxa)
}
\description{
Generate the line and ribbon plot for the dashboard.
}
\examples{
data(test_data)
make_line_plot(test_maindata, "Group_5", c("hormonal proteins","structural proteins","enzymes","storage proteins","antibodies","transport proteins"))

}
