% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_line_plot.R
\docType{data}
\name{make_line_plot}
\alias{make_line_plot}
\title{make line plot}
\format{
A numeric vector.
}
\source{
ggplot2::geom_ribbon
}
\usage{
make_line_plot(dfgroup_long, selected_groups, selected_taxa)
}
\description{
Generate the line and ribbon plot for the dashboard.
}
\examples{
data(FuncExample)
make_line_plot(dfgroup_long4, c("Group_1", "Group_10"), c("Bacteria", "No Label"))
}
