% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_line_plot.R
\docType{data}
\name{make_line_plot}
\alias{make_line_plot}
\title{Make line plot}
\usage{
make_line_plot(dfgroup_long, selected_groups, selected_taxa)
}
\description{
Generate the line and ribbon plot for the dashboard.
}
\details{
this function makes a ribbon plot for every brush action in the Shiny to show the min, max, and mean value of each clustered pattern group across time. The ribbon plot is grouped and colored by the \code{type} variable in input datasets.
See also \code{ggplot2::geom_ribbon}.
}
\examples{
data(test_data)
make_line_plot(test_maindata, "Group_5", c("hormonal proteins","structural proteins","enzymes","storage proteins","antibodies","transport proteins"))

}
