% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stackbar_plot.R
\docType{data}
\name{make_stackbar_plot}
\alias{make_stackbar_plot}
\title{make stackbar plot}
\format{
A numeric vector.
}
\source{
ggplot2::geom_bar
}
\usage{
make_stackbar_plot(dfgroup_long, selected_groups, selected_taxa)
}
\description{
Generate the stackbar plot for the dashboard.
}
\examples{
data(FuncExample)
make_stackbar_plot(dfgroup_long,c('Group_33','Group_3'),c('Bacteria',"Metazoa","Burkholderia"))
}
