% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_the_graph.R
\docType{package}
\name{make_the_graph}
\alias{make_the_graph}
\title{Make graph plot}
\source{
ggraph::ggraph
}
\usage{
make_the_graph(ptw, network_output, min_weight, s_ptw)
}
\description{
Generate the graph for the dashboard.
}
\examples{
data(test_data)
make_the_graph(test_graphptw, test_network, 0.03, "Muscular System")

}
