% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_the_graph.R
\docType{data}
\name{make_the_graph}
\alias{make_the_graph}
\title{make graph plot
need more edit on doc}
\format{
A numeric vector. pathway$Pathway
}
\source{
ggplot2::geom_bar
}
\usage{
make_the_graph(ptw, network_output, min_weight, s_ptw)
}
\description{
Generate the graph for the dashboard.
}
\examples{
data(FuncExample)
make_the_graph(ptw, networkres, 0.03, "03010_Ribosome")
}
