% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_produce_lm__.R
\docType{package}
\name{mass_produce_lm__}
\alias{mass_produce_lm__}
\title{Produce linear model for each column}
\usage{
mass_produce_lm__(data)
}
\arguments{
\item{data}{A dataframe.}
}
\description{
This is an internal function designed to automatically generate a list of linear functions for a dataset. Each column is considered a response variable against all other columns. The resulting functions can be utilized as inputs for random forest regression.
}
