% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_database.R
\docType{package}
\name{match_database}
\alias{match_database}
\title{Match database}
\usage{
match_database(data, id_colname, id_type)
}
\description{
\code{match_database()} is an internal function that matches the selected columns in the dataset with corresponding databases.
}
\examples{
data(test_data)
head(test_annotations_processed)
head(match_database(test_annotations_processed,id_colname = c("GO_ID","KEGG_ID"),id_type = c("GO","KEGG")))

}
