% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_URL.R
\docType{package}
\name{paste_URL}
\alias{paste_URL}
\title{paste_URL}
\format{
A dataframe and a weblink column name
}
\usage{
paste_URL(x, id_type)
}
\description{
Generate the cleaned version of the pathway for the dashboard.
}
\examples{
data(stuff)
nao_ptw <- ptw_process(ptw, ko_term)
head(nao_ptw)
}
