% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\docType{package}
\name{pre_process}
\alias{pre_process}
\title{Pre-processing datasets}
\usage{
pre_process(
  data,
  typenameList = NULL,
  replaceNA = TRUE,
  scale = TRUE,
  autoColName = "Sec_"
)
}
\arguments{
\item{data}{A data.frame to describe each feature in one row.
The data should contain variables \code{ID} \code{value on time_1}, ..., \code{value on time_k}, and \code{type} for extracting patterns across the time.
Note that the initial and last column must be exactly \code{ID} and \code{type}.
If multiple data.frame with the above format needs to be analyzed, you could also put a list of data.frame into this argument.
In this case, variable \code{type} is not required and will be generated by the next argument \code{typenameList}.}

\item{typenameList}{A vector of strings. This parameter is used to clarify the source or names for each data.frame, and is only applicable when the input of \code{data} is a list of data.frame. By default, it will be set as "Dataset_1", "Dataset_2", ..., etc.}

\item{scale}{Logical; if \code{scale} is TRUE (default), standardize the data.frame by row with \code{base::scale}. This converts each original value into a z-score. See also \code{scale_by_row__()}.}

\item{autoColName}{A string; if \code{autoColName} is not-NULL (default), it will automatically set uniform column names for all the data.frames.
This parameter is only applicable when the input of \code{data} is a list of data.frame.}

\item{replaceNa}{Logical; if \code{replaceNa} is TRUE (default), replace NA with 0.}
}
\value{
The function returns a long data.frame with columns \code{ID}, \code{value on time_1}, ..., \code{value on time_k}, and \code{type}.
}
\description{
The \code{pre_process()} function aids in processing data inputs and automatically establishes a standardized format for future use.
It allows for two types of data input: a list of datasets from different sources or a long dataset containing a specified last column \code{type}.
}
\details{
We consider two distinct scenarios for this application:
\itemize{
\item In one scenario, individuals collect several datasets from various aspects and instruments for the same objects. For example, they might be separately detecting lipids, metabolites, and peptides from a specific soil sample.
\item In the other scenario, all the data is of uniform quality, but it can be categorized into larger groups that exhibit significant differences.
In both of these cases, the pre_process() function serves as a valuable and versatile tool. Yet, this function is optional when generating the dashboard. Users can perform their own processing as long as the format matches the required output. However, they should be mindful that the number of samples (timepoints) must be greater than 5 to avoid potential errors in the subsequent prediction section.
}
}
\examples{
data(test_data)
head(test_data, 10)
a <- pre_process(test_data)
head(a, 10)

}
