% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_for_make_functions.R
\docType{package}
\name{reshape_for_make_functions}
\alias{reshape_for_make_functions}
\title{Reshape for 'make' functions}
\usage{
reshape_for_make_functions(data, cluster, annotation, id_colname, id_type)
}
\arguments{
\item{data}{The output of \code{pre_process()}}

\item{cluster}{The output of \code{gClusters()}}

\item{annotation}{The output of \code{gPathway()}}

\item{id_colname}{The columns that contain database IDs.}

\item{id_type}{The corresponding database names for the above columns.}
}
\description{
This internal function produces three primary datasets for the dashboard intended for the "make" functions.
}
\examples{
data(test_data)
l <- reshape_for_make_functions(test_data_processed, test_cluster, test_annotations_processed, id_colname = c("GO_ID","KEGG_ID"),id_type = c("GO","KEGG"))
head(l[[1]])
head(l[[2]])
head(l[[3]])


}
