% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_by_row__.R
\docType{package}
\name{scale_by_row__}
\alias{scale_by_row__}
\title{Scale by row}
\usage{
scale_by_row__(data)
}
\description{
Scales the values for each sample, where each row is independently processed.
}
\details{
The input is expected to be a data frame with the first column as the ID and the following columns containing observations at different time points. The ID column remains unaltered, and the other columns should be in double (dbl) format and will be scaled.
}
\examples{
data(test_data)
scale_by_row__(test_data[1:5,1:10])

}
